requirements.txt
// Acá ponemos las librerías


pandas
openpyxl
numpycd 